requirements.txt
fastapi
uvicorn[standard]
pydantic
python-docx
python-pptx